# api/openapi.yaml
openapi: 3.0.3
info:
  title: Fitness Tracker API
  version: 1.0.0
  description: |
    Lean Go backend for the core fitness tracking loop: users, exercise library,
    workouts, workout exercises, workout sets, meals, and weight entries.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh an existing session
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New access and refresh tokens.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        default:
          $ref: "#/components/responses/Error"
...
  /v1/auth/sessions/{id}:
    delete:
      tags: [Auth]
      summary: Revoke a specific session
      operationId: deleteSession
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    ID:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NoContent:
      description: Resource deleted successfully.
  schemas:
...
    ErrorResponse:
      type: object
      required: [title, status, error]
      properties:
        type:
          type: string
          description: Problem type URI. Validation errors use a dedicated type; generic errors use `about:blank`.
          example: about:blank
        title:
          type: string
          description: Short, human-readable summary of the error category.
          example: Bad Request
        status:
          type: integer
          description: HTTP status code for the error response.
          example: 400
        detail:
          type: string
          description: Human-readable explanation of the specific error instance.
          example: invalid email or password
        error:
          type: string
          description: Backward-compatible alias of `detail`.
          example: invalid email or password
        errors:
          type: object
          description: Field-specific validation messages when the request fails validation.
          additionalProperties:
            type: string
          example:
            email: email is required
            password: password must be at least 8 characters
...
    RefreshTokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT